Core: Updated to 11.1.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/f9bf082be962d45de79f936d625f644253e9b810)

+++ Source/Game/DataStorage/Structs/C_Records.cs
@@ -129,7 +129,7 @@ namespace Game.DataStorage
         public ushort CinematicSequenceID;
         public ushort DefaultSpec;
         public uint Id;
-        public byte PrimaryStatPriority;
+        public sbyte PrimaryStatPriority;
         public PowerType DisplayPower;
         public byte RangedAttackPowerPerAgility;
         public byte AttackPowerPerAgility;
@@ -532,7 +532,7 @@ namespace Game.DataStorage
         public uint FootstepCameraEffectID;
         public uint DeathThudCameraEffectID;
         public uint SoundID;
-        public uint SizeClass;
+        public sbyte SizeClass;
         public float CollisionWidth;
         public float CollisionHeight;
         public float WorldEffectScale;
